refactor: 정책 부분의 볼륨이 커짐에 따른 별도 패키지 분리
