+++ src/main/java/com/neuma573/autoboard/mvc/controller/MvcController.java
@@ -4,8 +4,8 @@ import com.neuma573.autoboard.board.model.annotation.CheckBoardAccess;
 import com.neuma573.autoboard.board.model.dto.BoardResponse;
 import com.neuma573.autoboard.board.model.enums.BoardAction;
 import com.neuma573.autoboard.board.service.BoardService;
-import com.neuma573.autoboard.global.service.PolicyService;
 import com.neuma573.autoboard.global.utils.RequestUtils;
+import com.neuma573.autoboard.policy.service.PolicyService;
 import com.neuma573.autoboard.post.model.annotation.CheckPostAccess;
 import com.neuma573.autoboard.post.model.dto.PostResponse;
 import com.neuma573.autoboard.post.model.enums.PostAction;
