+++ src/main/java/com/neuma573/autoboard/policy/service/PolicyService.java
@@ -1,8 +1,8 @@
-package com.neuma573.autoboard.global.service;
+package com.neuma573.autoboard.policy.service;
 
 import com.neuma573.autoboard.global.exception.AccessDeniedException;
-import com.neuma573.autoboard.global.model.dto.PolicyResponse;
-import com.neuma573.autoboard.global.repository.PolicyRepository;
+import com.neuma573.autoboard.policy.model.dto.PolicyResponse;
+import com.neuma573.autoboard.policy.repository.PolicyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
